refactor(services): shrink score_run function; split scoring package into raw-scores and project-scores modules
