+++ src/quodeq/services/score_run.py
@@ -47,6 +47,75 @@ def _read_project_source_file_count(reports_dir: str, project: str) -> int:
     return read_scan_total_files(Path(reports_dir) / project)
 
 
+def _score_one_dimension(
+    dim_id: str, jsonl_path: Path, run_id: str, evaluation_dir: Path,
+    source_file_count: int, files_read: int,
+    compiled_dir: Path | None, evaluators_dir: Path | None, params,
+    *, parser, scorer, reporter, log,
+) -> None:
+    """Score one dimension's evidence and write its report. Logs and
+    swallows any parse/score failure (fail-soft: one bad dimension must not
+    stop scoring the rest)."""
+    try:
+        evidence = parser(jsonl_path, EvidenceContext(
+            language="", repository="", date_str="",
+            source_file_count=source_file_count, files_read=files_read,
+        ), compiled_dir=compiled_dir, evaluators_dir=evaluators_dir,
+            req_map_reader=read_req_to_principle_map,
+            refs_reader=load_compiled_refs,
+            cwe_url_template=cwe_url_template(),
+            on_quarantine=log_quarantined_findings,
+            on_malformed_line=log_malformed_jsonl_line)
+        if evidence is None:
+            return
+        scores = scorer(evidence, mode="numerical", params=params)
+        reporter(evidence, scores, dim_id, evaluation_dir)
+        log.info(
+            "Scored cancelled dimension '%s' for run %s (files_read=%d)",
+            dim_id, run_id[:8], files_read,
+        )
+    except (OSError, json.JSONDecodeError, ValueError, KeyError) as exc:
+        log.debug("Could not score cancelled dimension '%s': %s", dim_id, exc)
+
+
+def _should_score_dimension(
+    dim_id: str, evidence_size: int, run_dir: Path, evaluation_dir: Path, dim_states: dict,
+) -> bool:
+    if dimension_report_exists(evaluation_dir, dim_id):
+        return False  # already scored
+    if dim_states.get(dim_id, {}).get("state") == "incomplete":
+        _logger.info("Skipping scoring for incomplete dim %s", dim_id)
+        return False
+    if evidence_size == 0:
+        return False  # no findings
+    # Only score dimensions that passed verification (analysis queue exists)
+    if not queue_file_exists(run_dir, dim_id):
+        return False  # verification not completed for this dimension
+    return True
+
+
+def _setup_scoring_context(reports_dir: str, project: str, run_dir: Path, evaluation_dir: Path):
+    """Resolve the scoring inputs shared across every dimension in this run.
+
+    Returns (evidence_entries, source_file_count, params, compiled_dir,
+    evaluators_dir, dim_states), or None when there's no evidence to score.
+    """
+    evidence_entries = list_dimension_evidence(run_dir)
+    if evidence_entries is None:
+        return None
+
+    ensure_dir(evaluation_dir)
+    source_file_count = _read_project_source_file_count(reports_dir, project)
+    params = load_params()
+    # Same standard dirs as a completed run's scoring, so off-standard
+    # findings are quarantined here too instead of entering the grade.
+    from quodeq.services.evidence_rescore import standard_dirs  # noqa: PLC0415
+    compiled_dir, evaluators_dir = standard_dirs()
+
+    dim_states = read_dimensions(run_dir).get("dimensions", {})
+    return evidence_entries, source_file_count, params, compiled_dir, evaluators_dir, dim_states
+
+
 def score_completed_evidence(
     reports_dir: str, job: dict,
     *,
@@ -74,54 +143,19 @@ def score_completed_evidence(
     if not project or not run_id:
         return
 
-    _log = _logger
-
     run_dir = Path(reports_dir) / project / run_id
     evaluation_dir = run_dir / "evaluation"
-    evidence_entries = list_dimension_evidence(run_dir)
-    if evidence_entries is None:
+    ctx = _setup_scoring_context(reports_dir, project, run_dir, evaluation_dir)
+    if ctx is None:
         return
-
-    ensure_dir(evaluation_dir)
-    source_file_count = _read_project_source_file_count(reports_dir, project)
-    params = load_params()
-    # Same standard dirs as a completed run's scoring, so off-standard
-    # findings are quarantined here too instead of entering the grade.
-    from quodeq.services.evidence_rescore import standard_dirs  # noqa: PLC0415
-    compiled_dir, evaluators_dir = standard_dirs()
-
-    dim_states = read_dimensions(run_dir).get("dimensions", {})
+    evidence_entries, source_file_count, params, compiled_dir, evaluators_dir, dim_states = ctx
 
     for dim_id, jsonl_path, evidence_size in evidence_entries:
-        if dimension_report_exists(evaluation_dir, dim_id):
-            continue  # already scored
-        if dim_states.get(dim_id, {}).get("state") == "incomplete":
-            _logger.info("Skipping scoring for incomplete dim %s", dim_id)
+        if not _should_score_dimension(dim_id, evidence_size, run_dir, evaluation_dir, dim_states):
             continue
-        if evidence_size == 0:
-            continue  # no findings
-        # Only score dimensions that passed verification (analysis queue exists)
-        if not queue_file_exists(run_dir, dim_id):
-            continue  # verification not completed for this dimension
-
         files_read = _read_queue_files_count(dimension_queue_file(run_dir, dim_id))
-        try:
-            evidence = parser(jsonl_path, EvidenceContext(
-                language="", repository="", date_str="",
-                source_file_count=source_file_count, files_read=files_read,
-            ), compiled_dir=compiled_dir, evaluators_dir=evaluators_dir,
-                req_map_reader=read_req_to_principle_map,
-                refs_reader=load_compiled_refs,
-                cwe_url_template=cwe_url_template(),
-                on_quarantine=log_quarantined_findings,
-                on_malformed_line=log_malformed_jsonl_line)
-            if evidence is None:
-                continue
-            scores = scorer(evidence, mode="numerical", params=params)
-            reporter(evidence, scores, dim_id, evaluation_dir)
-            _log.info(
-                "Scored cancelled dimension '%s' for run %s (files_read=%d)",
-                dim_id, run_id[:8], files_read,
-            )
-        except (OSError, json.JSONDecodeError, ValueError, KeyError) as exc:
-            _log.debug("Could not score cancelled dimension '%s': %s", dim_id, exc)
+        _score_one_dimension(
+            dim_id, jsonl_path, run_id, evaluation_dir, source_file_count, files_read,
+            compiled_dir, evaluators_dir, params,
+            parser=parser, scorer=scorer, reporter=reporter, log=_logger,
+        )
